+++ src/main/java/lab1/common/ShapeFactory.java
@@ -16,11 +16,11 @@ public class ShapeFactory {
         return new Triangle(vertex1, vertex2, vertex3, outlineColor, fillColor);
     }
 
-    public static IShape createRectangle(Point leftTop, double width, double height, Color outlineColor, Color fillColor) {
+    public static IShape createRectangle(Point leftTop, int width, int height, Color outlineColor, Color fillColor) {
         return new Rectangle(leftTop, width, height, outlineColor, fillColor);
     }
 
-    public static IShape createCircle(Point center, double radius, Color outlineColor, Color fillColor) {
+    public static IShape createCircle(Point center, int radius, Color outlineColor, Color fillColor) {
         return new Circle(center, radius, outlineColor, fillColor);
     }
 }
